Move monitoring tests to test_monitor.py
